utility: refactor tls memory allocator
remove c interface of tls memory allocator, and move
related code to utility module
