+++ include/dsn/tool-api/task.h
@@ -38,13 +38,13 @@
 
 #include <dsn/utility/ports.h>
 #include <dsn/utility/extensible_object.h>
+#include <dsn/utility/callocator.h>
+#include <dsn/utility/utils.h>
+#include <dsn/utility/binary_writer.h>
 #include <dsn/tool-api/task_spec.h>
 #include <dsn/tool-api/task_tracker.h>
 #include <dsn/tool-api/rpc_message.h>
-#include <dsn/cpp/callocator.h>
 #include <dsn/tool-api/auto_codes.h>
-#include <dsn/utility/utils.h>
-#include <dsn/utility/binary_writer.h>
 
 namespace dsn {
